+++ modules/cms/server/lib/dust/block.js
@@ -40,7 +40,7 @@ module.exports = (dust) => {
         injectedChunk.end();
       }, (error) => {
         dust.helpers.error(injectedChunk, error);
-        console.log(error, error.stack);
+        console.error(error, error.stack);
         injectedChunk.end();
       });
     });